pyproject.toml
[project]
name = "hotel-management"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "django>=5.2.5",
    "djangorestframework>=3.16.1",
]
